Modify minigame timers to use global config

+++ Assets/Scripts/Minigames/DireDodging/States/DireDodgingGameplayState.cs
@@ -20,7 +20,7 @@ public class DireDodgingGameplayState : IDireDodgingState {
         pauseService = ServiceLocatorAccessor.GetService<IPauseService>();
     }
 
-    private void OnHalfwayPointReached(int remainingTimeInSeconds) {
+    private void OnHalfwayPointReached(float remainingTimeInSeconds) {
         timer.OnHalfwayPointReached -= OnHalfwayPointReached;
         DireDodgingMinigameManager.Instance.StartIncreasingIntensity(remainingTimeInSeconds);
         DireDodgingMinigameManager.Instance.SetMusicIntensity(2);
